refactor(add-axios): refatora nome de pasta
