+++ src/context/TransactionContext.tsx
@@ -1,4 +1,5 @@
 import { ReactNode, createContext, useEffect, useState } from 'react'
+import { api } from '../lib/api'
 
 export interface TransactionType {
   id: number
@@ -24,16 +25,13 @@ export const TransactionsProvider = ({
   const [transactions, setTransactions] = useState<TransactionType[]>([])
 
   const fetchTransactions = async (query?: string) => {
-    const url = new URL('http://localhost:3000/transactions')
+    const response = await api.get('/transactions', {
+      params: {
+        q: query,
+      },
+    })
 
-    if (query) {
-      url.searchParams.append('q', query)
-    }
-
-    const response: Response = await fetch(url)
-    const data: TransactionType[] = await response.json()
-
-    setTransactions(data)
+    setTransactions(response.data)
   }
 
   useEffect(() => {
